% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Difference.R
\name{CompareProductVersions}
\alias{CompareProductVersions}
\title{Compare Product Versions}
\usage{
CompareProductVersions(product, version, previous = NULL)
}
\arguments{
\item{product}{A character value. The CDISC product}

\item{version}{A character value. The version of the CDISC version being used}

\item{previous}{A character value. A previous version to compare
the \code{version} to. If not supplied will compare with preceeding version.}
}
\value{
A character value. A sentence with a joke
}
\description{
Compare a product with a different version
}
\seealso{
\url{https://www.cdisc.org/cdisc-library/api-documentation}
}
\author{
Gareth Burns
}
